requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.4
matplotlib>=3.7
seaborn>=0.13
plotly>=5.18
openpyxl>=3.1
pyyaml>=6.0
joblib>=1.3
streamlit>=1.32
jupyter>=1.0
ipykernel>=6.29
